requirements.txt
python-dotenv==0.18.0
requests==2.25.1
rich==10.4.0
sense-energy==0.9.0
